+++ src/EventListener/LoadDataContainerListener.php
@@ -20,6 +20,9 @@ class LoadDataContainerListener
             case 'tl_reader_config_element':
                 $this->addConfigElementFields($table);
                 return;
+            case 'tl_list_config':
+                $this->addListConfigDcaFields();
+                return;
         }
     }
 
@@ -38,5 +41,37 @@ class LoadDataContainerListener
                 ['huh.tiny_slider.backend.tiny_slider_spread', 'editTinySliderConfig'],
             ],
         ];
+    }
+
+    /**
+     * Add fields to tl_list_config if list bundle is installed
+     */
+    protected function addListConfigDcaFields()
+    {
+        $dca = &$GLOBALS['TL_DCA']['tl_list_config'];
+        $dca['palettes']['default'] = str_replace('addMasonry', 'addMasonry,addTinySlider', $dca['palettes']['default']);
+        $dca['palettes']['__selector__'][] = 'addTinySlider';
+        $dca['subpalettes']['addTinySlider'] = 'tinySliderConfig';
+
+        $fields = [
+            'addTinySlider'    => [
+                'label'     => &$GLOBALS['TL_LANG']['tl_list_config']['addTinySlider'],
+                'exclude'   => true,
+                'inputType' => 'checkbox',
+                'eval'      => ['tl_class' => 'w50 clr', 'submitOnChange' => true],
+                'sql'       => "char(1) NOT NULL default ''",
+            ],
+            'tinySliderConfig' => [
+                'label'            => &$GLOBALS['TL_LANG']['tl_list_config']['tinySliderConfig'],
+                'exclude'          => true,
+                'filter'           => true,
+                'inputType'        => 'select',
+                'options_callback' => ['huh.tiny_slider.backend.tiny_slider_spread', 'getBaseConfigurations'],
+                'eval'             => ['tl_class' => 'w50', 'mandatory' => true, 'includeBlankOption' => true, 'submitOnChange' => true],
+                'sql'              => "int(10) unsigned NOT NULL default '0'",
+            ],
+        ];
+
+        $dca['fields'] = array_merge($fields, is_array($dca['fields']) ? $dca['fields'] : []);
     }
 }