+++ src/Resources/contao/languages/de/tl_list_config.php
@@ -1,0 +1,11 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_list_config'];
+
+/**
+ * Fields
+ */
+$lang['addTinySlider'][0]    = 'Als Tiny-Slider ausgeben';
+$lang['addTinySlider'][1]    = 'Wählen Sie diese Option, um die Liste als Tiny-Slider zu rendern.';
+$lang['tinySliderConfig'][0] = 'Tiny-Slider Konfiguration';
+$lang['tinySliderConfig'][1] = 'Wählen Sie hier die gewünschte Tiny-Slider Konfiguration aus.';