+++ src/Resources/contao/languages/en/tl_list_config.php
@@ -1,0 +1,11 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_list_config'];
+
+/**
+ * Fields
+ */
+$lang['addTinySlider'][0]    = 'Render as Tiny slider';
+$lang['addTinySlider'][1]    = 'Select this option to render the list as a tiny slider.';
+$lang['tinySliderConfig'][0] = 'Tiny slider configuration';
+$lang['tinySliderConfig'][1] = 'Select the desired tiny slider configuration here.';