+++ webpack.config.js
@@ -2,16 +2,19 @@ var Encore = require('@symfony/webpack-encore');
 
 Encore
 .setOutputPath('src/Resources/public/')
-.addEntry('contao-tiny-slider-bundle', './src/Resources/assets/js/contao-tiny-slider-bundle.js')
+.addEntry('contao-tiny-slider-bundle', './src/Resources/npm-package/js/contao-tiny-slider-bundle.js')
 .addEntry('contao-tiny-slider-bundle-theme', './src/Resources/assets/js/contao-tiny-slider-bundle-theme.js')
-.addEntry('tiny-slider', 'tiny-slider/dist/tiny-slider')
 .setPublicPath('/bundles/contaotinyslider/')
 .setManifestKeyPrefix('bundles/contaotinyslider')
 .disableSingleRuntimeChunk()
-.addExternals({
-    'tiny-slider': 'tns'
+.splitEntryChunks()
+    .configureSplitChunks(function(splitChunks) {
+        splitChunks.name =  function (module, chunks, cacheGroupKey) {
+            const moduleFileName = module.identifier().split('/').reduceRight(item => item).split('.').slice(0, -1).join('.');
+            return `${moduleFileName}`;
+        };
 })
-.configureBabel()
+.configureBabel(null)
 .enableSourceMaps(!Encore.isProduction())
 .enableSassLoader()
 .enablePostCssLoader()